requirements.txt
python-telegram-bot==20.7
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
